# backend/storage/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: 7KAIH MVP API - Revised Validation Model
  version: 2.0.0
  description: |
    API MVP untuk sistem 7 Kebiasaan Anak Indonesia Hebat (7KAIH).

    Revisi utama versi 2.0:
    - Validasi dilakukan pada level item kebiasaan, bukan hanya daily check-in.
    - Siswa tidak perlu mengirim konteks rumah/sekolah saat check-in.
    - Orang tua/wali memvalidasi item anak berdasarkan relasi aktif student_parent_relations.
    - Guru memvalidasi item siswa berdasarkan kelas yang dia ampu.
    - Satu item check-in dapat memiliki validasi terpisah dari orang tua dan guru.
    - Orang tua/wali juga dapat memonitor perkembangan anaknya.
    - Guru tetap menjadi aktor utama untuk monitoring kelas dan rekap.
servers:
  - url: /api/v1
    description: Base API v1
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Login user
      description: Digunakan oleh semua role.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginSuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout user
      responses:
        '200':
          description: Logout berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicNullResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags: [Auth]
      summary: Ambil data user login
      responses:
        '200':
          description: Data user berhasil diambil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /profile:
    get:
      tags: [Auth]
      summary: Ambil profil user login
      description: Digunakan oleh semua role untuk melihat data profil sendiri.
      responses:
        '200':
          description: Profil berhasil diambil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [Auth]
      summary: Update profil user login
      description: Digunakan oleh semua role untuk mengubah data profil sendiri.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profil berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /profile/password:
    put:
      tags: [Auth]
      summary: Update password user login
      description: Digunakan oleh semua role untuk mengubah password akun sendiri.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordRequest'
      responses:
        '200':
          description: Password berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicNullResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /roles:
    get:
      tags: [Master Data]
      summary: Ambil daftar role
      responses:
        '200':
          description: Daftar role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /habits:
    get:
      tags: [Master Data]
      summary: Ambil daftar 7 kebiasaan aktif
      responses:
        '200':
          description: Daftar habits
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HabitsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /classes:
    get:
      tags: [Master Data]
      summary: Ambil daftar kelas
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: is_active
          schema:
            type: integer
            enum: [0, 1]
      responses:
        '200':
          description: Daftar kelas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /student/checkins/today:
    get:
      tags: [Student]
      summary: Ambil form check-in hari ini
      description: |
        Hanya role siswa. Mengambil daftar habits aktif, default context,
        dan check-in hari ini jika sudah ada.
      responses:
        '200':
          description: Data form check-in hari ini
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentTodayCheckinResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /student/checkins:
    get:
      tags: [Student]
      summary: Ambil riwayat check-in siswa login
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        '200':
          description: Riwayat check-in siswa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentCheckinHistoryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Student]
      summary: Create atau update daily check-in
      description: |
        Hanya role siswa. Bersifat upsert.
        Satu siswa hanya boleh memiliki satu check-in per tanggal.
        Untuk MVP, checkin_date kosong berarti tanggal hari ini.
        Jika checkin_date dikirim, backend wajib memastikan tanggal tersebut adalah hari ini.

        Jika siswa mengubah item yang sudah tervalidasi:
        - perubahan notes tidak membatalkan validasi;
        - perubahan is_done pada item dapat membatalkan atau memerlukan ulang validasi item tersebut.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentCheckinUpsertRequest'
      responses:
        '200':
          description: Check-in berhasil disimpan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentCheckinDetailSaveResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /student/checkins/{id}:
    get:
      tags: [Student]
      summary: Ambil detail check-in siswa login
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Detail check-in siswa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentCheckinDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /student/recap:
    get:
      tags: [Student]
      summary: Ambil rekap personal siswa login
      parameters:
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        '200':
          description: Rekap personal siswa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentRecapResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /parent/children:
    get:
      tags: [Parent]
      summary: Ambil daftar anak orang tua login
      description: Hanya role orang_tua. Mengambil anak yang punya relasi aktif.
      responses:
        '200':
          description: Daftar anak
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParentChildrenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /parent/children/{studentId}/checkins:
    get:
      tags: [Parent]
      summary: Ambil daftar check-in anak
      description: Orang tua hanya boleh melihat anak yang punya relasi aktif.
      parameters:
        - $ref: '#/components/parameters/StudentIdPath'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
        - in: query
          name: validated
          schema:
            type: integer
            enum: [0, 1]
      responses:
        '200':
          description: Daftar check-in anak
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParentChildCheckinListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /parent/checkins/{id}:
    get:
      tags: [Parent]
      summary: Ambil detail check-in anak
      description: Orang tua hanya boleh melihat check-in anak yang punya relasi aktif.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Detail check-in anak
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParentCheckinDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /parent/checkins/{id}/validate-home:
    post:
      tags: [Parent]
      summary: Validasi batch item check-in oleh orang tua
      description: |
        Hanya role orang_tua. Endpoint ini memvalidasi item yang:
        - milik anak dengan relasi aktif;
        - is_done = true.

        Backend menyimpan validasi dengan validator_role = orang_tua dan validation_source = rumah.
        Jika item_ids tidak dikirim, backend memvalidasi semua item is_done = true pada check-in tersebut.
        Endpoint idempotent per kombinasi daily_checkin_item_id + validator_role.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateBatchRequest'
      responses:
        '200':
          description: Validasi rumah berhasil atau sudah tervalidasi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateBatchResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /parent/checkin-items/{id}/validate:
    post:
      tags: [Parent]
      summary: Validasi satu item check-in oleh orang tua
      description: |
        Hanya role orang_tua. Item harus is_done = true.
        Orang tua harus punya relasi aktif dengan siswa pemilik item.
        Backend menyimpan validasi dengan validator_role = orang_tua dan validation_source = rumah.
        Endpoint idempotent per kombinasi daily_checkin_item_id + validator_role.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Item berhasil divalidasi atau sudah tervalidasi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateItemResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /parent/children/{studentId}/recap:
    get:
      tags: [Parent]
      summary: Ambil rekap anak untuk orang tua
      parameters:
        - $ref: '#/components/parameters/StudentIdPath'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        '200':
          description: Rekap anak
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentRecapResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /teacher/classes:
    get:
      tags: [Teacher]
      summary: Ambil daftar kelas guru login
      responses:
        '200':
          description: Daftar kelas guru
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherClassesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /teacher/classes/{classId}/students:
    get:
      tags: [Teacher]
      summary: Ambil daftar siswa berdasarkan kelas
      description: Guru hanya boleh mengakses kelas yang dia ampu.
      parameters:
        - $ref: '#/components/parameters/ClassIdPath'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          description: Daftar siswa per kelas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherClassStudentsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /teacher/classes/{classId}/checkins:
    get:
      tags: [Teacher]
      summary: Monitoring check-in per kelas
      description: Guru hanya boleh mengakses kelas yang dia ampu.
      parameters:
        - $ref: '#/components/parameters/ClassIdPath'
        - in: query
          name: date
          schema:
            type: string
            format: date
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: checked_in
          schema:
            type: integer
            enum: [0, 1]
        - in: query
          name: validated
          schema:
            type: integer
            enum: [0, 1]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Monitoring check-in per kelas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherClassCheckinsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /teacher/checkins/{id}:
    get:
      tags: [Teacher]
      summary: Ambil detail check-in siswa untuk guru
      description: Guru hanya boleh melihat check-in siswa dari kelas yang dia ampu.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Detail check-in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherCheckinDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /teacher/checkins/{id}/validate-school:
    post:
      tags: [Teacher]
      summary: Validasi batch item check-in oleh guru
      description: |
        Hanya role guru. Endpoint ini memvalidasi item yang:
        - milik siswa pada kelas yang dia ampu;
        - is_done = true.

        Backend menyimpan validasi dengan validator_role = guru dan validation_source = sekolah.
        Jika item_ids tidak dikirim, backend memvalidasi semua item is_done = true pada check-in tersebut.
        Endpoint idempotent per kombinasi daily_checkin_item_id + validator_role.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateBatchRequest'
      responses:
        '200':
          description: Validasi sekolah berhasil atau sudah tervalidasi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateBatchResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /teacher/checkin-items/{id}/validate:
    post:
      tags: [Teacher]
      summary: Validasi satu item check-in oleh guru
      description: |
        Hanya role guru. Item harus is_done = true.
        Guru harus mengampu kelas siswa pemilik item.
        Backend menyimpan validasi dengan validator_role = guru dan validation_source = sekolah.
        Endpoint idempotent per kombinasi daily_checkin_item_id + validator_role.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Item berhasil divalidasi atau sudah tervalidasi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateItemResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /teacher/classes/{classId}/weekly-recap:
    get:
      tags: [Teacher]
      summary: Ambil rekap mingguan per kelas
      parameters:
        - $ref: '#/components/parameters/ClassIdPath'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        '200':
          description: Rekap mingguan kelas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherPeriodRecapResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /teacher/classes/{classId}/monthly-recap:
    get:
      tags: [Teacher]
      summary: Ambil rekap bulanan per kelas
      description: Mendukung status sudah_terbiasa atau belum_terbiasa dengan rule 2/3 hari periode.
      parameters:
        - $ref: '#/components/parameters/ClassIdPath'
        - in: query
          name: month
          required: true
          description: Format YYYY-MM, contoh 2026-04.
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
      responses:
        '200':
          description: Rekap bulanan kelas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherPeriodRecapResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /teacher/students/{studentId}/recap:
    get:
      tags: [Teacher]
      summary: Ambil rekap detail siswa untuk guru
      description: Guru hanya boleh mengakses siswa dari kelas yang dia ampu.
      parameters:
        - $ref: '#/components/parameters/StudentIdPath'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        '200':
          description: Rekap siswa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentRecapResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/dashboard-summary:
    get:
      tags: [Admin]
      summary: Ambil ringkasan dashboard admin
      description: Ringkasan statistik utama untuk halaman Dashboard Admin.
      responses:
        '200':
          description: Ringkasan dashboard admin berhasil diambil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDashboardSummaryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/users:
    get:
      tags: [Admin]
      summary: Ambil daftar user
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - in: query
          name: role
          schema:
            type: string
            enum: [siswa, orang_tua, guru, admin]
        - in: query
          name: class_id
          schema:
            type: integer
            format: int64
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: is_active
          schema:
            type: integer
            enum: [0, 1]
      responses:
        '200':
          description: Daftar user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUsersListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Admin]
      summary: Buat user baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreateUserRequest'
      responses:
        '201':
          description: User berhasil dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /admin/users/{id}:
    get:
      tags: [Admin]
      summary: Ambil detail user
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Detail user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Admin]
      summary: Update user
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateUserRequest'
      responses:
        '200':
          description: User berhasil diupdate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Admin]
      summary: Soft delete user
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: User berhasil dihapus
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicNullResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/users/{id}/reset-password:
    post:
      tags: [Admin]
      summary: Reset password user
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminResetPasswordRequest'
      responses:
        '200':
          description: Password berhasil direset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicNullResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseSimple'
...
  schemas:
...
    LoginRequest:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          example: budi123
        password:
          type: string
          format: password
          example: secret123
    LoginSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Login successful
        data:
          type: object
          properties:
            access_token:
              type: string
              example: jwt-or-token
            token_type:
              type: string
              example: Bearer
            expires_in:
              type: integer
              example: 86400
            user:
              $ref: '#/components/schemas/UserSummary'
...